landing-zone-ai/python-ai/requirements.txt
flask==3.0.0
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
opencv-python==4.8.0.76
Pillow==10.0.0
matplotlib==3.8.0
